feat: Scrape game metadata universally
Introduce a more general-purpose scraper function that prioritizes OpenGraph and Twitter Card meta tags for title extraction. This allows the scraper to work with a wider range of game URLs beyond just AN1.com. The default console is set to 'Android' for APKs, with the expectation of later user editing.